Refactor StreamUnpackerReader
