+++ src/MsgPack/StreamUnpackerReader.TypedRead.tt
@@ -34,7 +34,13 @@
 #endif
 
 using System;
+#if CORE_CLR || UNITY || NETSTANDARD1_1
+using Contract = MsgPack.MPContract;
+#else
+using System.Diagnostics.Contracts;
+#endif // CORE_CLR || UNITY || NETSTANDARD1_1
 using System.IO;
+using System.Text;
 #if FEATURE_TAP
 using System.Threading;
 using System.Threading.Tasks;
@@ -86,7 +92,11 @@ foreach ( var type in scalarTypes )
 		}
 <#
 	}
-}
+} // foreach
+
+this.WriteRead( /* isAsync */false );
+this.WriteReadString( /* isAsync */false );
+this.WriteDrain( /* isAsync */false );
 #>
 
 #if FEATURE_TAP
@@ -129,7 +139,11 @@ foreach ( var type in scalarTypes )
 		}
 <#
 	}
-}
+}// foreach
+
+this.WriteRead( /* isAsync */true );
+this.WriteReadString( /* isAsync */true );
+this.WriteDrain( /* isAsync */true );
 #>
 
 #endif // FEATURE_TAP
@@ -203,4 +217,204 @@ private void WriteStreamReaderReadBody( string type, string primitiveType, bool 
 			}
 <#+
 } // WriteStreamReaderReadBody
+
+private void WriteRead( bool isAsync )
+{
+#>
+
+		// TODO: Use Span<T>
+		public override <#= isAsync ? "async Task" : "void" #> Read<#= isAsync ? "Async" : String.Empty #>( byte[] buffer, int size<#= isAsync ? ", CancellationToken cancellationToken" : String.Empty #> )
+		{
+#if DEBUG
+			if ( this._source.CanSeek )
+			{
+				Contract.Assert( this._source.Position == this._offset, this._source.Position + "==" + this._offset );
+			}
+#endif // DEBUG
+
+			// Reading 0 byte from stream causes exception in some implementation (issue #60, reported from @odyth).
+			if ( size == 0 )
+			{
+				return;
+			}
+
+			this._lastOffset = this._offset;
+			var remaining = size;
+			var offset = 0;
+			int read;
+
+			do
+			{
+<#+
+	if ( !isAsync )
+	{
+#>
+				read = this._source.Read( buffer, offset, remaining );
+<#+
+	}
+	else
+	{
+#>
+				read = await this._source.ReadAsync( buffer, offset, remaining, cancellationToken ).ConfigureAwait( false );
+<#+
+	}
+#>
+				remaining -= read;
+				offset += read;
+			} while ( read > 0 && remaining > 0 );
+
+			this._offset += offset;
+#if DEBUG
+			if ( this._source.CanSeek )
+			{
+				Contract.Assert( this._source.Position == this._offset, this._source.Position + "==" + this._offset );
+			}
+#endif // DEBUG
+
+			if ( offset < size )
+			{
+				this.ThrowEofException( size );
+			}
+		}
+<#+
+} // WriteRead
+
+
+private void WriteReadString( bool isAsync )
+{
+#>
+
+		public override <#= isAsync ? "async Task<string>" : "string" #> ReadString<#= isAsync ? "Async" : String.Empty #>( int length<#= isAsync ? ", CancellationToken cancellationToken" : String.Empty #> )
+		{
+			// Reading 0 byte from stream causes exception in some implementation (issue #60, reported from @odyth).
+			if ( length == 0 )
+			{
+				return String.Empty;
+			}
+
+			var bytes = BufferManager.NewByteBuffer( length );
+
+			if ( length <= bytes.Length )
+			{
+<#+
+	if ( !isAsync )
+	{
+#>
+				this.Read( bytes, length );
+<#+
+	}
+	else
+	{
+#>
+				await this.ReadAsync( bytes, length, cancellationToken ).ConfigureAwait( false );
+<#+
+	}
+#>
+				return Encoding.UTF8.GetString( bytes, 0, length );
+			}
+
+			var decoder = Encoding.UTF8.GetDecoder();
+			var chars = BufferManager.NewCharBuffer( bytes.Length );
+			var stringBuffer = new StringBuilder( length );
+			var remaining = length;
+			bool isCompleted;
+			do
+			{
+				var reading = Math.Min( remaining, bytes.Length );
+				this._lastOffset = this._offset;
+<#+
+	if ( !isAsync )
+	{
+#>
+				var bytesRead = this._source.Read( bytes, 0, reading );
+<#+
+	}
+	else
+	{
+#>
+				var bytesRead = await this._source.ReadAsync( bytes, 0, reading, cancellationToken ).ConfigureAwait( false );
+<#+
+	}
+#>
+				this._offset += bytesRead;
+				if ( bytesRead == 0 )
+				{
+					this.ThrowEofException( reading );
+				}
+
+				isCompleted = decoder.DecodeString( bytes, 0, bytesRead, chars, stringBuffer );
+				remaining -= bytesRead;
+			} while ( remaining > 0 );
+
+			if ( !isCompleted )
+			{
+				this.ThrowBadUtf8Exception();
+			}
+
+			return stringBuffer.ToString();
+		}
+<#+
+}
+
+private void WriteDrain( bool isAsync )
+{
+#>
+
+		public override <#= isAsync ? "async Task<bool>" : "bool" #> Drain<#= isAsync ? "Async" : String.Empty #>( uint size<#= isAsync ? ", CancellationToken cancellationToken" : String.Empty #> )
+		{
+			// Reading 0 byte from stream causes exception in some implementation (issue #60, reported from @odyth).
+			if ( size == 0 )
+			{
+				return true;
+			}
+
+			if ( this._useStreamPosition )
+			{
+				var remaining = this._source.Length - this._source.Position;
+				if ( remaining >= size )
+				{
+					this._source.Position += size;
+					this._offset += size;
+					return true;
+				}
+				else
+				{
+					return false;
+				}
+			}
+			else
+			{
+				var dummyBufferForSkipping = BufferManager.NewByteBuffer( unchecked(( int )Math.Min( size, Int32.MaxValue )) );
+				long remaining = size;
+				while ( remaining > 0 )
+				{
+					var reading = unchecked( ( int )Math.Min( remaining, dummyBufferForSkipping.Length ) );
+					this._lastOffset = this._offset;
+<#+
+	if ( !isAsync )
+	{
+#>
+					var lastRead = this._source.Read( dummyBufferForSkipping, 0, reading );
+<#+
+	}
+	else
+	{
+#>
+					var lastRead = await this._source.ReadAsync( dummyBufferForSkipping, 0, reading, cancellationToken ).ConfigureAwait( false );
+<#+
+	}
+#>
+					this._offset += lastRead;
+					remaining -= lastRead;
+					if ( lastRead == 0 )
+					{
+						return false;
+					}
+				}
+
+				return true;
+			}
+		}
+<#+
+}
 #>
